Revert "ie15: convert to a device with frontends (standalone driver and rs232…"
