+++ scripts/src/bus.lua
@@ -1764,8 +1764,6 @@ if (BUSES["RS232"]~=null) then
 		MAME_DIR .. "src/devices/bus/rs232/terminal.h",
 		MAME_DIR .. "src/devices/bus/rs232/xvd701.cpp",
 		MAME_DIR .. "src/devices/bus/rs232/xvd701.h",
-		MAME_DIR .. "src/devices/bus/rs232/ie15.cpp",
-		MAME_DIR .. "src/devices/bus/rs232/ie15.h",
 	}
 end
